+++ garage/envs/__init__.py
@@ -3,6 +3,8 @@ from garage.envs.base import Step
 from garage.envs.env_spec import EnvSpec
 from garage.envs.grid_world_env import GridWorldEnv
 from garage.envs.identification_env import IdentificationEnv  # noqa: I100
+from garage.envs.noisy_env import DelayedActionEnv
+from garage.envs.noisy_env import NoisyObservationEnv
 from garage.envs.normalized_env import normalize
 from garage.envs.sliding_mem_env import SlidingMemEnv
 
@@ -12,6 +14,8 @@ __all__ = [
     "EnvSpec",
     "GridWorldEnv",
     "IdentificationEnv",
+    "DelayedActionEnv",
+    "NoisyObservationEnv",
     "normalize",
     "SlidingMemEnv",
 ]
