+++ garage/envs/mujoco/randomization/__init__.py
@@ -1,4 +1,6 @@
 from garage.envs.mujoco.randomization.randomized_env import randomize
+from garage.envs.mujoco.randomization.variation import Distribution
+from garage.envs.mujoco.randomization.variation import Method
 from garage.envs.mujoco.randomization.variation import Variations
 
-__all__ = ["randomize", "Variations"]
+__all__ = ["randomize", "Distribution", "Method", "Variations"]
