+++ garage/sampler/__init__.py
@@ -1,3 +1,6 @@
+from garage.sampler.base import BaseSampler
+from garage.sampler.base import Sampler
+from garage.sampler.batch_sampler import BatchSampler
 from garage.sampler.stateful_pool import singleton_pool
 
-__all__ = ["singleton_pool"]
+__all__ = ["BaseSampler", "BatchSampler", "Sampler", "singleton_pool"]
