+++ garage/tf/core/__init__.py
@@ -1,11 +1,7 @@
-from garage.tf.core import layers
 from garage.tf.core.parameterized import Parameterized
 from garage.tf.core.layers_powered import LayersPowered  # noqa: I100
-from garage.tf.core.network import MLP
 
 __all__ = [
-    "layers",
     "Parameterized",
     "LayersPowered",
-    "MLP",
 ]
