+++ garage/tf/policies/deterministic_mlp_policy.py
@@ -3,8 +3,8 @@ import tensorflow as tf
 from garage.core import Serializable
 from garage.misc.overrides import overrides
 from garage.tf.core import LayersPowered
-from garage.tf.core import MLP
 import garage.tf.core.layers as L
+from garage.tf.core.network import MLP
 from garage.tf.misc import tensor_utils
 from garage.tf.policies import Policy
 from garage.tf.spaces import Box
