+++ garage/tf/policies/gaussian_mlp_policy.py
@@ -5,8 +5,8 @@ from garage.core import Serializable
 from garage.misc import logger
 from garage.misc.overrides import overrides
 from garage.tf.core import LayersPowered
-from garage.tf.core import MLP
 import garage.tf.core.layers as L
+from garage.tf.core.network import MLP
 from garage.tf.distributions import DiagonalGaussian
 from garage.tf.misc import tensor_utils
 from garage.tf.policies import StochasticPolicy
