+++ garage/theano/core/__init__.py
@@ -1,4 +1,3 @@
 from garage.theano.core.lasagne_powered import LasagnePowered
-from garage.theano.core.network import MLP
 
-__all__ = ["LasagnePowered", "MLP"]
+__all__ = ["LasagnePowered"]
