+++ garage/theano/regressors/categorical_mlp_regressor.py
@@ -7,7 +7,7 @@ import theano.tensor as TT
 from garage.core import Serializable
 from garage.misc import logger
 from garage.theano.core import LasagnePowered
-from garage.theano.core import MLP
+from garage.theano.core.network import MLP
 from garage.theano.distributions import Categorical
 from garage.theano.misc import tensor_utils
 from garage.theano.optimizers import LbfgsOptimizer
